#2423 Ajout tests
