+++ base/models/learning_unit_component_class.py
@@ -50,8 +50,8 @@ class LearningUnitComponentClass(models.Model):
         )
 
 
-def find_by_id(learning_unit_component_class_id):
-    return LearningUnitComponentClass.objects.get(pk=learning_unit_component_class_id)
+def find_by_id(learning_unit_component_class):
+    return LearningUnitComponentClass.objects.get(pk=learning_unit_component_class.id)
 
 
 def find_by_learning_class_year(learning_class_year):
